avoid crash in rare situations
(cherry picked from commit f5d7ef36c9bb7e6b45efcb16e0b061619784012c)

+++ app/models.py
@@ -2213,7 +2213,7 @@ class Post(db.Model):
         # Handle file deletions from disk before cascade deletes the File records
         if self.image_id and self.image:
             self.image.delete_from_disk(purge_cdn=False)
-        if self.type == POST_TYPE_VIDEO and _store_files_in_s3() and self.url.startswith(f'https://{current_app.config["S3_PUBLIC_URL"]}'):
+        if self.type == POST_TYPE_VIDEO and _store_files_in_s3() and self.url and self.url.startswith(f'https://{current_app.config["S3_PUBLIC_URL"]}'):
             from app.shared.tasks.maintenance import delete_from_s3
             if current_app.debug:
                 delete_from_s3([self.url])
